+++ project-base/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - admin: E-mail newsletter: now contains list of registered e-mails with ability to delete them
 - Frontend site for requesting GDPR information (@stanoMilan) 
     - Admin - added mail template personal data request
+    - Frontend send email with link personal information  
 
 ### Changed
 - cache is cleared before PHPUnit tests only when run via [Phing targets](docs/introduction/phing-targets.md), not when run using `phpunit` directly (@PetrHeinz)